Respect configured measurement setup and relocate video export helpers
