+++ Utility/Exports/VideoExportProgressReport.cs
@@ -1,3 +1,3 @@
-namespace ElectricalImpedanceTomography.ViewModels;
+namespace Utility.Exports;
 
 public sealed record VideoExportProgressReport(double Progress, string StatusMessage);
